fix: CI&Release 7

+++ core/src/mingwX64Main/kotlin/com/adoktl/util/PlatformMingw.kt
@@ -2,4 +2,4 @@ package com.adoktl.util
 
 import platform.windows.GetTickCount64
 
-actual fun currentTimeMillis(): Long = GetTickCount64()
+actual fun currentTimeMillis(): Long = GetTickCount64().toLong()
